% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/assign.groups.for.batched.phyloscanner.analysis.R
\name{assign.groups.for.batched.phyloscanner.analysis}
\alias{assign.groups.for.batched.phyloscanner.analysis}
\title{Group individuals for batched phyloscanner analysis}
\usage{
assign.groups.for.batched.phyloscanner.analysis(x, batch.size = 50)
}
\arguments{
\item{x}{Character vector of individual identifiers.}

\item{batch.size}{Batch size. Default is 50.}
}
\value{
tibble with rows 'IND' (individual identifiers), 'PTY_RUN' group for phyloscanner analysis, and 'BATCH' batch of individuals (not used further, but there should be two batches of individuals in each phyloscanner analysis).
}
\description{
This function groups individuals for phyloscanner analyses, so that phylogenetic linkage between every pair of individuals is assessed at least once.
  Specifically, individuals are grouped into batches of specified size, and then, all possible pairs of batches are formed. 
  Each of these pairs of batches defines a group of individuals between whom phylogenetic linkages are assessed in one phyloscanner run.
  The number of individuals in each group is twice the batch size.
}
\examples{
x <- c("15-01402","15-04719","16-00616","16-00801","16-01173","16-01191","16-01302","16-01408","16-01414","16-01465","16-01581","16-01663","16-03259","16-03499","16-03516","16-03644","16-03807")
pty.runs	<- phyloscannerR::assign.groups.for.batched.phyloscanner.analysis(x, batch.size=50)
}
\author{
Oliver Ratmann
}
