% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_summary_methods.R
\name{classify.pairwise.relationships}
\alias{classify.pairwise.relationships}
\title{Classify pairwise host relationships in deep sequence phylogenies Make viral phylogenetic classification}
\usage{
classify.pairwise.relationships(ptrees, close.threshold = 0.025,
  distant.threshold = 0.05, relationship.types = c("proximity.3.way",
  "any.ancestry", "close.x.contiguous", "close.and.contiguous",
  "close.and.adjacent", "close.and.contiguous.and.directed",
  "close.and.adjacent.and.directed",
  "close.and.contiguous.and.ancestry.cat",
  "close.and.adjacent.and.ancestry.cat"), verbose = FALSE)
}
\arguments{
\item{ptrees}{A list of class \code{phyloscanner.trees} produced by \code{phyloscanner.analyse.trees}.}

\item{close.threshold}{The (potentially normalised) patristic threshold used to determine if two patients' subgraphs are "close".}

\item{distant.threshold}{If present, a second distance threshold determines hosts that are "distant" from each other, with those lying between \code{close.threshold} and \code{dist.threshold} classed as "intermediate". The default is the same as \code{close.threshold}, so the intermediate class does not exist.}

\item{relationship.types}{Classification types.
\itemize{
 \item{"proximity.3.way"}{Classify individuals by phylogenetic distance between subgraphs. Suggested use: to exclude phylogenetic linkage based on distance alone.}
 \item{"close.and.contiguous"}{Classify individuals by phylogenetic distance and contiguity of subgraphs. Suggested use: to identify phylogenetically linked pairs.}
 \item{"close.and.adjacent"}{Classify individuals by phylogenetic distance and adjacency of subgraphs. Suggested use: to identify phylogenetically linked pairs.}
 \item{"close.and.contiguous.and.directed"}{Classify ancestry among contiguous subgraphs. Suggested use: to identify direction of transmission based on contiguous subgraphs.}
 \item{"close.and.adjacent.and.directed"}{Classify ancestry among adjacent subgraphs. Suggested use: to identify direction of transmission based on adjacent subgraphs.} 
 \item{"close.and.contiguous.and.ancestry.cat"}{Classify contiguity and ancestry between individuals. Suggested use: to determine probabilities for transmission networks.}
 \item{"close.and.adjacent.and.ancestry.cat"}{Classify adjacency and ancestry between individuals. Suggested use: to determine probabilities for transmission networks.}
}}

\item{verbose}{Verbose output}
}
\value{
A data frame with viral phylogenetic classifications of pairwise host relationships in each deep sequence phylogeny
}
\description{
Classify pairwise host relationships in deep sequence phylogenies Make viral phylogenetic classification
}
\examples{
\dontrun{
require(phyloscannerR)
#
# Example on data from Rakai Community Cohort Study
# load phyloscanner output from 'phyloscanner.analyse.trees'
#	
file	<- system.file(file.path('extdata','ptyr192_phsc_analyse_trees_output.RData'),package='phyloscannerR')
load(file)	#loads 'phsc', output from 'phyloscanner.analyse.trees'
#	use distance thresholds found in analysis of Rakai couples
close.threshold <- 0.025
distant.threshold <- 0.05	
#	use relationship types based on adjacency
#	this also considers linkage etc between individuals who have dual infections, recombinants etc 
#	..and thus may not have *all* their subgraphs adjacent to each other
relationship.types <- c('close.and.adjacent',					
		'close.and.adjacent.and.directed',					
		'close.and.adjacent.and.ancestry.cat')	
dwin <- classify.pairwise.relationships(phsc, 
           close.threshold=close.threshold, 
           distant.threshold=distant.threshold,
           relationship.types=relationship.types, 
           verbose=TRUE)
}	
}
\author{
Oliver Ratmann, Matthew Hall
}
