% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_line_utilities.R
\name{cmd.phyloscanner.analyse.trees.valid.args}
\alias{cmd.phyloscanner.analyse.trees.valid.args}
\title{Obtain valid input arguments for a phyloscanner analysis on a tree or set of trees}
\usage{
cmd.phyloscanner.analyse.trees.valid.args(prog.phyloscanner_analyse_trees)
}
\arguments{
\item{prog.phyloscanner_analyse_trees}{The full file name of \code{phyloscanner_analyse_trees.R}.}
}
\description{
Obtain valid input arguments for a phyloscanner analysis on a tree or set of trees
}
\seealso{
\link{\code{cmd.phyloscanner.analyse.trees}}
}
