% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.pairwise.relationships.R
\name{count.pairwise.relationships}
\alias{count.pairwise.relationships}
\title{Count pairwise relationships across deep-sequence trees}
\usage{
count.pairwise.relationships(dwin, w.slide = NA, verbose = TRUE)
}
\arguments{
\item{dwin}{A data frame produced by \code{classify.pairwise.relationships}.}

\item{w.slide}{Increment between genomic windows. Default: NA.}

\item{verbose}{Verbose output. Default: TRUE.}
}
\value{
A data frame with counts of viral phylogenetic classifications between pairs of individuals.
}
\description{
Count pairwise relationships across deep-sequence trees
}
\examples{
\dontrun{
require(phyloscannerR)
#
#	continue Rakai example,
#	load phyloscanner output from 'phyloscanner.analyse.trees'
#	
file	<- system.file(file.path('extdata','ptyr192_phsc_analyse_trees_output.R'),package='phyloscannerR')
load(file)	#loads 'phsc', output from 'phyloscanner.analyse.trees'
#	use distance thresholds found in analysis of Rakai couples
close.threshold <- 0.025
distant.threshold <- 0.05	
#	use relationship types based on adjacency
#	this also considers linkage etc between individuals who have dual infections, recombinants etc 
#	..and thus may not have *all* their subgraphs adjacent to each other
relationship.types <- c('proximity.3.way',					
		'close.and.adjacent',					
		'close.and.adjacent.and.directed',					
		'close.and.adjacent.and.ancestry.cat')	
dwin <- classify.pairwise.relationships(phsc, allow.mt=TRUE, close.threshold=close.threshold, distant.threshold=distant.threshold,relationship.types=relationship.types, verbose=TRUE)
tip.regex <- "^(.*)_fq[0-9]+_read_([0-9]+)_count_([0-9]+)$"
min.reads <- 30
min.tips <- 1
dwin <- select.windows.by.read.and.tip.count(phsc, dwin, tip.regex, min.reads, min.tips)
# count phylogenetic relationships across deep-sequence trees
dc <- count.pairwise.relationships(dwin)
#
# 	end of Rakai example
#
}	
}
\author{
Oliver Ratmann, Matthew Hall
}
