% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_analysis.R
\name{draw.summary.statistics}
\alias{draw.summary.statistics}
\title{Graph summary statistics for a single host}
\usage{
draw.summary.statistics(phyloscanner.trees, sum.stats, host, verbose = F)
}
\arguments{
\item{phyloscanner.trees}{A list of class \code{phyloscanner.trees}}

\item{sum.stats}{The output of a call to \code{gather.summary.statistics}.}

\item{host}{The host to obtain graphs for.}

\item{verbose}{Verbose output}
}
\description{
Graph summary statistics for a single host
}
