% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utility_functions.R
\name{extract.subgraph}
\alias{extract.subgraph}
\title{Extract subgraph from phyloscanner tree}
\usage{
extract.subgraph(ph, mrca)
}
\arguments{
\item{ph}{A tree of class \code{SIMMAP} and \code{phyloscanner}, i.e. with elements \code{maps}, \code{mapped.edge}, \code{node.states} and with attributes \code{SPLIT}, \code{INDIVIDUAL}, \code{BRANCH_COLOURS}, \code{SUBGRAPH_MRCA}.}

\item{mrca}{Node index of subgraph MRCA}
}
\value{
Subgraph in ape format, with additional elements \code{subgraph.name} (subgraph name in the phyloscanner SPLIT attribute), \code{subgraph.root.edge} (length of the ancestral edge of the subgraph MRCA), \code{subgraph.parent.state} (ancestral state of the parent of subgraph MRCA).
}
\description{
Extract subgraph from phyloscanner tree
}
\examples{
#
# Example on data from Rakai Community Cohort Study
#
require(phyloscannerR)
ph.phyloscanner <- readRDS(file=system.file(file.path('extdata','ptyr292_1600_to_1849.rds'),package='phyloscannerR'))

# add SIMMAP elements to tree
# because internally the extract.subgraph function uses functions written for
# trees in SIMMAP format
ph.phsc.plus.simmap <- phyloscanner.to.simmap(ph.phyloscanner, delete.phyloscanner.structures=FALSE)

# extract subgraph MRCAS and choose all those corresponding to a particular host
host <- 'RkA07714M'
mrcas <- which( attr(ph.phsc.plus.simmap, 'SUBGRAPH_MRCA') )
mrcas <- mrcas[ attr(ph.phsc.plus.simmap, 'INDIVIDUAL')[mrcas]==host ]	

# extract subgraphs
subgraphs <- lapply(mrcas, function(mrca) extract.subgraph(ph.phsc.plus.simmap, mrca))
# the first subgraph consists of just 1 taxon and is therefore hard to represent as an ape object
# note the dummy structures
# the second subgraph consists of 9 taxa and is in standard ape format
# note the extra elements "subgraph.name", "subgraph.root.edge", "subgraph.parent.state"
str( subgraphs[[2]] )
}
\seealso{
\code{\link{phyloscanner.to.simmap}}
}
\author{
Oliver Ratmann
}
