% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_network_functions.R
\name{find.most.likely.chains.RBGLedmonds}
\alias{find.most.likely.chains.RBGLedmonds}
\title{Find most likely transmission chains}
\usage{
find.most.likely.chains.RBGLedmonds(rtnn, verbose = 0)
}
\arguments{
\item{rtnn}{tibble encoding the three edges of a phyloscanner transmission network. Must have columns 'H1','H2','IDCLU','TYPE','SCORE','K_EFF'.}
}
\value{
tibble encoding the most likely chain. Has columns 'H1','H2','IDCLU', 'LINK_12', 'LINK_21' (either 1 or 0 for a link in the corresponding direction), and 'MX_PROB_12', 'MX_PROB_21' (associated posterior probabilities)
}
\description{
This function finds the most likely transmission 
chain among all chains spanning the nodes in a specified transmission network.
The transmission network consists of at most three edges between a set of
individuals (directed edge in either direction, and undirected edge). Chains 
are defined as spanning graphs through the set of nodes, without loops and with
in-degree equal to one for all nodes, except the start node. Each directed edge 
in a chain has a weight, which corresponds to the phylogenetic evidence of transmission
in this direction. It is set to the phyloscanner score for transmission in this 
direction plus half the phyloscanner score of the undirected edge, for transmission 
with direction unclear. The probability of the entire chain is given by the product
of the phyloscanner scores along each edge in the chain.   
Edmonds algorithm is used to solve for the most probable chain.
}
\seealso{
\code{\link{find.networks}}
}
\author{
Oliver Ratmann
}
\keyword{internal}
