% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_network_functions.R
\name{find.networks}
\alias{find.networks}
\title{Find phylogenetic transmission networks and most likely transmission chain}
\usage{
find.networks(
  dc,
  control = list(linked.group = "close.and.adjacent.cat", linked.no =
    "not.close.or.nonadjacent", linked.yes = "close.and.adjacent", dir.group =
    "close.and.adjacent.and.directed.cat", neff.cut = 3, weight.complex.or.no.ancestry =
    0.5),
  verbose = TRUE
)
}
\arguments{
\item{dc}{Summary of phylogenetic relationship counts for each pair, stored as tibble.}

\item{control}{List of control variables:
\itemize{
        \item{\code{linked.group}} Phyloscanner classification used to identify pairs in networks. Default 'close.and.adjacent.cat'.
        \item{\code{linked.no}} Phyloscanner classification type quantifying that pairs are not linked. Default 'not.close.or.nonadjacent'.
        \item{\code{linked.yes}} Phyloscanner classification type quantifying that pairs are linked. Default 'close.and.adjacent'.
        \item{\code{neff.cut}} Threshold on the minimum number of deep-sequence phylogenies with sufficient reads from two individuals to make any phylogenetic inferences. Default: 3.
        \item{\code{weight.complex.or.no.ancestry}} Weight given to score complex.or.no.ancestry. Default: 50%.
}}

\item{verbose}{Flag to switch on/off verbose mode. Default: TRUE.}
}
\value{
list of two R objects
\itemize{
        \item{\code{transmission.networks}} is a tibble that describes the edge list of pairs of individuals in a network, and corresponding phyloscanner scores
        \item{\code{most.likely.transmission.chains}} is a tibble that describes the edge list of pairs of individuals in the most likely chain, and corresponding phyloscanner scores
}
See description.
}
\description{
This function computes transmission networks from phyloscanner output of a population-based deep sequence sample.
A transmission network is defined as a set of individuals between whom phylogenetic linkage is not excluded.
Every individual in the network has at least one partner in the network between whom evidence for being phylogenetically unlinked is below a threshold.
These pairs of individuals are identified with a separate function, \code{\link{find.pairs.in.networks}}.
Due to the nature of the deep-sequence
data, there are up to three edges between pairs of individuals, giving the strength of evidence
of spread in each direction (two possibilities) and the strength of evidence for phylogenetic linkage with
the direction remaining unclear. Some of these pairs have limited evidence for phylogenetic linkage. The networks
are best interpreted as partially sampled transmission chain.

This function also finds the most likely transmission
chain among all chains spanning the nodes in a specified transmission network.
The transmission network consists of at most three edges between a set of
individuals (directed edge in either direction, and undirected edge). Chains
are defined as spanning graphs through the set of nodes, without loops and with
in-degree equal to one for all nodes, except the start node. Each directed edge
in a chain has a weight, which corresponds to the phylogenetic evidence of transmission
in this direction. It is set to the phyloscanner score for transmission in this
direction plus half the phyloscanner score of the undirected edge, for transmission
with direction unclear. The probability of the entire chain is given by the product
of the phyloscanner scores along each edge in the chain.
}
\seealso{
\code{\link{find.pairs.in.networks}}, \code{\link{plot.network}}, \code{\link{plot.chain}}
}
\author{
Oliver Ratmann
}
