% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_network_functions.R
\name{find.pairs.in.networks}
\alias{find.pairs.in.networks}
\title{Find pairs of individuals between whom linkage is not excluded phylogenetically}
\usage{
find.pairs.in.networks(
  dwin,
  dc,
  control = list(linked.group = "close.and.adjacent.cat", linked.no =
    "not.close.or.nonadjacent", linked.yes = "close.and.adjacent", conf.cut = 0.6,
    neff.cut = 3),
  dmeta = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dwin}{A \code{data.frame} describing pairwise relationships between the hosts in each tree; normally output of \code{classify.pairwise.relationships}}

\item{dc}{A \code{data.frame} summarising pairwise relationships between the hosts across all trees; normally output of \code{count.pairwise.relationships}}

\item{control}{List of control variables:
\itemize{
        \item{\code{linked.group}} Phyloscanner classification used to identify pairs in networks. Default 'close.and.adjacent.cat'.
        \item{\code{linked.no}} Phyloscanner classification type quantifying that pairs are not linked. Default 'not.close.or.nonadjacent'.
        \item{\code{linked.yes}} Phyloscanner classification type quantifying that pairs are linked. Default 'close.and.adjacent'.
        \item{\code{conf.cut}} Threshold on the proportion of deep-sequence phylogenies with distant/disconnected subgraphs above which pairs are considered phylogenetically unlinked. Default: 0.6
        \item{\code{neff.cut}} Threshold on the minimum number of deep-sequence phylogenies with sufficient reads from two individuals to make any phylogenetic inferences. Default: 3.
}}

\item{dmeta}{Optional individual-level meta-data that is to be added to output. Can be NULL.}

\item{verbose}{Flag to switch on/off verbose mode. Default: TRUE.}
}
\value{
Three R objects are generated:
\itemize{
        \item{\code{network.pairs}} is a tibble that describes pairs of individuals between whom linkage is not excluded phylogenetically.
        \item{\code{relationship.counts}} is a tibble that summarises the phylogenetic relationship counts for each pair.
        \item{\code{windows}} is a tibble that describes the basic phyloscanner statistics (distance, adjacency, paths between subgraphs) in each deep-sequence phylogeny for each pair.
}
}
\description{
This function identifies pairs of individuals between whom linkage is not excluded phylogenetically in a large number of phyloscanner analyses, and provides detailed information on them.
}
\seealso{
\code{\link{phyloscanner.analyse.trees}}, \code{\link{cmd.phyloscanner.analyse.trees}}
}
\author{
Oliver Ratmann
}
