% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_functions.R
\name{get.keff.and.neff}
\alias{get.keff.and.neff}
\title{Count observed relationship states}
\usage{
get.keff.and.neff(df, relationship.types, w.slide = NA)
}
\arguments{
\item{df}{tibble with basic relationship types for paired individuals across windows. Must contain columns 'ID1','ID2','W_FROM','W_TO','TYPE_BASIC'.}

\item{relationship.types}{names of relationship groups}
}
\value{
new tibble with columns host.1 host.2 GROUP TYPE K KEFF N NEFF.
}
\description{
This function counts for each pair of individuals their relationship states across all windows (KEFF), and the total number of windows (NEFF). Since windows can be overlapping, the count is a real value.
}
\keyword{internal}
