% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.pairwise.relationships.R
\name{get.pairwise.relationships.basic.by.paths}
\alias{get.pairwise.relationships.basic.by.paths}
\title{Define basic topological relationships between two individuals by paths}
\usage{
get.pairwise.relationships.basic.by.paths(all.classifications)
}
\description{
Define basic topological relationships between two individuals by paths
}
\author{
Oliver Ratmann
}
\keyword{internal}
