% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.pairwise.relationships.R
\name{get.pairwise.relationships.basic}
\alias{get.pairwise.relationships.basic}
\title{Define basic topological relationships between two individuals}
\usage{
get.pairwise.relationships.basic(all.classifications)
}
\description{
Define basic topological relationships between two individuals
}
\author{
Oliver Ratmann
}
\keyword{internal}
