% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_summary_methods.R
\name{get.pairwise.relationships.close.and.adjacent.and.ancestry.cat}
\alias{get.pairwise.relationships.close.and.adjacent.and.ancestry.cat}
\title{Group to determine probabilities for transmission networks based on adjacent subgraphs}
\usage{

  get.pairwise.relationships.close.and.adjacent.and.ancestry.cat(all.classifications)
}
\description{
Define basic topological relationships 'close and 12', 'close and 21', 'close and ambiguous direction', 'not close or not adjacent'
}
\author{
Oliver Ratmann, Matthew Hall
}
\keyword{internal}
