% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.pairwise.relationships.R
\name{get.pairwise.relationships.close.and.adjacent.and.directed.cat}
\alias{get.pairwise.relationships.close.and.adjacent.and.directed.cat}
\title{Group to determine direction of transmission in likely pairs based on adjacent subgraphs}
\usage{

  get.pairwise.relationships.close.and.adjacent.and.directed.cat(all.classifications)
}
\description{
Group to determine direction of transmission in likely pairs based on adjacent subgraphs
}
\author{
Oliver Ratmann, Matthew Hall
}
\keyword{internal}
