% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.pairwise.relationships.R
\name{get.pairwise.relationships.close.and.adjacent.cat}
\alias{get.pairwise.relationships.close.and.adjacent.cat}
\title{Group to determine phylogenetic linkage based on adjacent subgraphs}
\usage{
get.pairwise.relationships.close.and.adjacent.cat(all.classifications)
}
\description{
Group to determine phylogenetic linkage based on adjacent subgraphs
}
\author{
Oliver Ratmann, Matthew Hall
}
\keyword{internal}
