% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.pairwise.relationships.R
\name{get.pairwise.relationships.proximity.3.way.cat}
\alias{get.pairwise.relationships.proximity.3.way.cat}
\title{Group to determine phylogenetic linkage based on distance only}
\usage{
get.pairwise.relationships.proximity.3.way.cat(all.classifications)
}
\description{
Group to determine phylogenetic linkage based on distance only
}
\author{
Oliver Ratmann, Matthew Hall
}
\keyword{internal}
