% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_functions.R
\name{get.pairwise.relationships}
\alias{get.pairwise.relationships}
\title{Calculate pairwise relationships}
\usage{
get.pairwise.relationships(df, relationship.types = c("proximity.3.way",
  "any.ancestry", "close.x.contiguous", "close.and.contiguous",
  "close.and.contiguous.and.directed", "close.and.adjacent.and.directed",
  "close.and.contiguous.and.ancestry.cat",
  "close.and.adjacent.and.ancestry.cat", "adjacent.and.proximity.cat"))
}
\arguments{
\item{df}{tibble to which new columns of relationship groups will be added. Must contain a column with name TYPE_BASIC. This column contains fundamental relationship states for every window, from which other relationships are derived.}

\item{relationship.types}{names of relationship groups}
}
\value{
input tibble with new columns. Each new column defines relationship states for a specific relationship group.
}
\description{
This function calculates pairwise relationships between each pair of two individuals in any window. Several different relationship groups can be calculated, for example just using pairwise distance, or using both pairwise distance and topology to define likely pairs.
}
\keyword{internal}
