% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_functions.R
\name{get.posterior.scores}
\alias{get.posterior.scores}
\title{Calculate marginal posterior probability for two individuals being in a particular relationship state}
\usage{
get.posterior.scores(df)
}
\arguments{
\item{df}{Input tibble}

\item{n.type}{Calibration parameter for the prior: minimum number of windows of state to select a pair of individuals with confidence of at least at least confidence.cut, if the total number of windows is n.obs}

\item{n.obs}{Calibration parameter for the prior: total number of windows.}

\item{confidence.cut}{Calibration parameter for the prior: confidence cut off.}
}
\value{
Input tibble with two additional columns POSTERIOR_ALPHA and POSTERIOR_BETA
}
\description{
This function calculates the parameters that specify the marginal posterior probability for two individuals being in a particular relationship state. The marginal posterior is Beta distributed and this function calculates the ALPHA and BETA parameters.
}
\keyword{internal}
