% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_functions.R
\name{multinomial.calculations}
\alias{multinomial.calculations}
\title{Calculate parameters of the posterior density for pairwise host relationships}
\usage{
multinomial.calculations(ptrees, close.threshold,
  tip.regex = "^(.*)_read_([0-9]+)_count_([0-9]+)$", allow.mt = F,
  min.reads = 0, min.tips = 0, distant.threshold = close.threshold,
  relationship.types = c("proximity.3.way", "any.ancestry",
  "close.x.contiguous", "close.and.contiguous",
  "close.and.contiguous.and.directed", "close.and.adjacent.and.directed",
  "close.and.contiguous.and.ancestry.cat",
  "close.and.adjacent.and.ancestry.cat", "adjacent.and.proximity.cat"),
  verbose = F)
}
\arguments{
\item{ptrees}{A list of class \code{phyloscanner.trees} produced by \code{phyloscanner.analyse.trees}.}

\item{close.threshold}{The (potentially normalised) patristic threshold used to determine if two patients' subgraphs are "close"}

\item{tip.regex}{The regular expression used to identify host IDs in tip names}

\item{allow.mt}{If FALSE, directionality is only inferred between pairs of hosts where a single clade from one host is nested in one from the other; this is more conservative.}

\item{min.reads}{The minimum number of reads from a host in a window needed in order for that window to count in determining relationships involving that patient}

\item{min.tips}{The minimum number of tips from a host in a window needed in order for that window to count in determining relationships involving that patient}

\item{distant.threshold}{If present, a second distance threshold determines hosts that are "distant" from each other, with those lying between \code{close.threshold} and \code{dist.threshold} classed as "intermediate". The default is the same as \code{close.threshold}, so the intermediate class does not exist.}

\item{verbose}{Verbose output}
}
\value{
A list with two items: \code{dwin} giving information on the genome windows for each pair of hosts, and \code{rplkl} giving information on phylogenetic relationships between each pair of hosts.
}
