% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_analysis.R
\name{pairwise.relationship.statistics}
\alias{pairwise.relationship.statistics}
\title{Draw bar graphs of pairwise topological/distance relationships to file}
\usage{
pairwise.relationship.statistics(ptrees,
  hosts = combn(all.hosts.from.trees(phyloscanner.trees), 2, simplify =
  F), file.name, height = 11.6929, width = 8.26772, verbose = F)
}
\arguments{
\item{ptrees}{A list of class \code{phyloscanner.trees}}

\item{hosts}{A list of pairs of hosts (as vectors) to obtain graphs for. By default, all pairs hosts detected in \code{ptrees}.}

\item{file.name}{Output file name (expected to be a PDF)}

\item{height}{The height of each page of the output file in inches (defaults to A4 size)}

\item{width}{The width of each page of the output file in inches (defaults to A4 size)}

\item{verbose}{Verbose output}
}
\description{
Draw bar graphs of pairwise topological/distance relationships to file
}
