% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_functions.R
\name{phsc.get.pairwise.relationships.keff.and.neff}
\alias{phsc.get.pairwise.relationships.keff.and.neff}
\title{Count observed relationship states}
\usage{
phsc.get.pairwise.relationships.keff.and.neff(df, get.groups, w.slide = NA)
}
\arguments{
\item{df}{data.table with basic relationship types for paired individuals across windows. Must contain columns 'ID1','ID2','W_FROM','W_TO','TYPE_BASIC'.}

\item{get.groups}{names of relationship groups}
}
\value{
new data.table with columns ID1 ID2 GROUP TYPE K KEFF N NEFF.
}
\description{
This function counts for each pair of individuals their relationship states across all windows (KEFF), and the total number of windows (NEFF). Since windows can be overlapping, the count is a real value.
}
\keyword{internal}
