% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_functions.R
\name{phsc.get.prior.parameter.n0}
\alias{phsc.get.prior.parameter.n0}
\title{Calculate prior parameter n0}
\usage{
phsc.get.prior.parameter.n0(n.states, keff = 2, neff = 3,
  confidence.cut = 0.66)
}
\arguments{
\item{n.states}{Number of relationship states.}

\item{confidence.cut}{Confidence cut off.}

\item{n.obs}{Minimum number of windows to select a pair of individuals with confidence at least confidence.cut.}
}
\value{
Prior parameter n0
}
\description{
This function calculates the prior parameter n0 such that the minimum number of windows needed is at least n.obs in order to select a pair of individuals with confidence at least confidence.cut.
}
\keyword{internal}
