% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_analysis.R
\name{phyloscanner.analyse.tree}
\alias{phyloscanner.analyse.tree}
\title{Perform a phyloscanner analysis on a single tree}
\usage{
phyloscanner.analyse.tree(tree.file.name, splits.rule = c("s", "r", "f"),
  sankoff.k = 0, sankoff.unassigned.switch.threshold = 0,
  continuation.unassigned.proximity.cost = 1000, outgroup.name = NULL,
  multifurcation.threshold = -1, guess.multifurcation.threshold = F,
  user.blacklist.file.name = NULL, duplicate.file.name = NULL,
  recombination.file.name = NULL,
  tip.regex = "^(.*)_read_([0-9]+)_count_([0-9]+)$",
  file.name.regex = "^\\\\D*([0-9]+)_to_([0-9]+)\\\\D*$",
  seed = sample(1:1e+07, 1), norm.ref.file.name = NULL,
  norm.standardise.gag.pol = F, norm.constants = NULL,
  parsimony.blacklist.k = 0, raw.blacklist.threshold = 0,
  ratio.blacklist.threshold = 0, do.dual.blacklisting = F,
  max.reads.per.host = Inf, blacklist.underrepresented = F, use.ff = F,
  prune.blacklist = F, read.counts.matter.on.zero.length.tips = T,
  verbosity = 0, no.progress.bars = F)
}
\arguments{
\item{tree.file.name}{The name of the tree file (Newick or NEXUS format).}

\item{splits.rule}{The rules by which the sets of hosts are split into groups in order to ensure that all groups can be members of connected subgraphs without causing conflicts. Options: s=Sankoff with optional within-host diversity penalty (slow, rigorous, recommended), r=Romero-Severson (quick, less rigorous with >2 hosts), f=Sankoff with continuation costs (experimental).}

\item{sankoff.k}{For \code{splits.rule = s} or \code{f} only. The \emph{k} parameter in the Sankoff reconstruction, representing the within-host diversity penalty.}

\item{sankoff.unassigned.switch.threshold}{For \code{splits.rule = s} only. Threshold at which a lineage reconstructed as infecting a host will transition to the unassigned state, if it would be equally parsimonious to remain in that host.}

\item{continuation.unassigned.proximity.cost}{For \code{splits.rule = f} only. The branch length at which an node is reconstructed as unassigned if all its neighbouring nodes are a greater distance away. The default is 1000, intended to be effectively infinite, such a node will never normally receive the unassigned state.}

\item{outgroup.name}{The name of the tip in the phylogeny/phylogenies to be used as outgroup (if unspecified, trees will be assumed to be already rooted). This should be sufficiently distant to any sequence obtained from a host that it can be assumed that the MRCA of the entire tree was not a lineage present in any sampled individual.}

\item{multifurcation.threshold}{If specified, branches shorter than this in the input tree will be collapsed to form multifurcating internal nodes. This is recommended; many phylogenetics packages output binary trees with short or zero-length branches indicating multifurcations.}

\item{guess.multifurcation.threshold}{Whether to guess the multifurcation threshold from the branch lengths of the trees and the width of the genomic window (if that information is available). It is recommended that trees are examined by eye to check that they do appear to have multifurcations if using this option.}

\item{user.blacklist.file.name}{The path of a text file containing the user-specified list of tips to be blacklisted}

\item{duplicate.file.name}{The path of a .csv file specifying which tree tips are from duplicate reads. Normally this is produced by \code{phyloscanner_make_trees.py}.}

\item{recombination.file.name}{The path for file containing the results of the \code{phyloscanner_make_trees.py} recombination metric analysis.}

\item{tip.regex}{Regular expression identifying tips from the dataset. This expects up to three capture groups, for host ID, read ID, and read count (in that order). If the latter two groups are missing then read information will not be used. The default matches input from the phyloscanner pipeline where the host ID is the BAM file name.}

\item{file.name.regex}{Regular expression identifying window coordinates. Two capture groups: start and end; if the latter is missing then the first group is a single numerical identifier for the window. The default matches input from the phyloscanner pipeline.}

\item{seed}{Random number seed; used by the downsampling process, and also ties in some parsimony reconstructions can be broken randomly.}

\item{norm.ref.file.name}{Name of a file giving a normalisation constant for every genome position. Cannot be used simultaneously with \code{norm.constants}. If neither is given then no normalisation will be performed.}

\item{norm.standardise.gag.pol}{Use only if \code{norm.ref.file.name} is given. An HIV-specific option: if true, the normalising constants are standardised so that the average on gag+pol equals 1. Otherwise they are standardised so the average on the whole genome equals 1.}

\item{norm.constants}{Either the path of a CSV file listing the file name for each tree (column 1) and the respective normalisation constant (column 2) or a single numerical normalisation constant to be applied to every tree. Cannot be used simultaneously with \code{norm.ref.file.name}. If neither is given then no normalisation will be performed.}

\item{parsimony.blacklist.k}{The \emph{k} parameter of the single-host Sankhoff parsimony reconstruction used to identify probable contaminants. A value of 0 is equivalent to not performing parsimony blacklisting.}

\item{raw.blacklist.threshold}{Used to specify a read count to be used as a raw threshold for duplicate or parsimony blacklisting. Use with \code{parsimony.blacklist.k} or \code{duplicate.file.regex} or both. Parsimony blacklisting will blacklist any subgraph with a read count strictly less than this threshold. Duplicate blacklisting will black list any duplicate read with a count strictly less than this threshold. The default value of 0 means nothing is blacklisted.}

\item{ratio.blacklist.threshold}{Used to specify a read count ratio (between 0 and 1) to be used as a threshold for duplicate or parsimony blacklisting. Use with \code{parsimony.blacklist.k} or \code{duplicate.file.regex} or both. Parsimony blacklisting will blacklist a subgraph if the ratio of its read count to the total read count from the same host is strictly less than this threshold. Duplcate blacklisting will blacklist a duplicate read if the ratio of its count to the count of the duplicate (from another host) is strictly less than this threshold.}

\item{do.dual.blacklisting}{Blacklist all reads from the minor subgraphs for all hosts established as dual by parsimony blacklisting (which must have been done for this to do anything).}

\item{max.reads.per.host}{Used to turn on downsampling. If given, reads will be blacklisted such that read counts (or tip counts if no read counts are identified) from each host are equal (although see \code{blacklist.underrepresented}.}

\item{blacklist.underrepresented}{If TRUE and \code{max.reads.per.host} is given, blacklist hosts from trees where their total tip count does not reach the maximum.}

\item{use.ff}{Use the \code{ff} package to store parsimony reconstruction matrices. Use if you run out of memory.}

\item{prune.blacklist}{If TRUE, all blacklisted and reference tips (except the outgroup) are pruned away before starting parsimony-based reconstruction.}

\item{read.counts.matter.on.zero.length.tips}{If TRUE, read counts on tips will be taken into account in parsimony reconstructions at the parents of zero-length terminal branches. Not applicable for the Romero-Severson-like reconstruction method.}

\item{verbosity}{The type of verbose output. 0=none, 1=minimal, 2=complete}

\item{no.progress.bars}{Hide the progress bars from verbose output.}
}
\value{
A list of class \code{phyloscanner.trees} with a single item of class \code{phyloscanner.tree}.
}
\description{
This function performs a parsimony reconstruction and classification of pairwise host relationships.
}
