% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{produce.pairwise.graphs2}
\alias{produce.pairwise.graphs2}
\title{Draw bar graphs of pairwise topological/distance relationships, version 2}
\usage{
produce.pairwise.graphs2(
  ptrees,
  hosts = all.hosts.from.trees(ptrees),
  inclusion = c("both", "either"),
  dwin = NULL,
  control = list(yintercept_close = 0.025, yintercept_dist = 1, breaks_x = seq(0,
    10000, 500), minor_breaks_x = seq(0, 10000, 100), breaks_y = c(0.001, 0.0025, 0.005,
    0.01, 0.025, 0.05, 0.1, 0.25), limits_y = c(0.001, 0.4), fill.topology = c(ancestral
    = "deepskyblue1", descendant = "deepskyblue4", intermingled = "#FDB863", sibling =
    "#8073AC", other = "grey80"))
)
}
\arguments{
\item{ptrees}{A list of class \code{phyloscanner.trees}}

\item{hosts}{A list of hosts (as a vector) to obtain graphs for. By default, all pairs of hosts detected in \code{ptrees}.}

\item{inclusion}{If "both", then only pairs in which both individuals are members of \code{hosts} are included. If "either" then pairs only need have one member from \code{hosts}}

\item{dwin}{Optional output of \code{classify.pairwise.relationships}. This can be specified to avoid double calculations.}

\item{control}{List of plotting attributes.}
}
\value{
A list whose elements are \code{data}, the underlying data frame for the graph, and \code{graph}, the graph itself.
}
\description{
This function generates scan plots that summarize reconstructed viral phylogenetic relationships of two individuals.
Several pairs of individuals can be processed simultaneously. For each pair of individuals, the scan plot shows the 
phylogenetic distance on the y-axis and topological relationship in colours between subgraphs from both individuals in each 
deep-sequence phylogeny across the genome. The genomic position on the x-axis indicates the start of each read alignment.
}
\examples{
#
# Example on data from Rakai Community Cohort Study
# remember that you can specify dwin to save computing time, if you have it already computed
#
\dontrun{	
file	<- system.file(file.path('extdata','ptyr192_phsc_analyse_trees_output.RData'),package='phyloscannerR')
load(file)	#loads 'phsc', output from 'phyloscanner.analyse.trees'	
hosts	<- c('RkA05868F','RkA05968M','RkA00369F','RkA01344M')
inclusion <- "both" 
tmp	<- produce.pairwise.graphs2(phsc, hosts=hosts, inclusion = "both")
tmp$graph
}
}
\seealso{
\link{\code{classify.pairwise.relationships}}
}
\author{
Oliver Ratmann
}
