% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestral_reconstruction_methods.R
\name{reconstruct.host.ancestral.sequences}
\alias{reconstruct.host.ancestral.sequences}
\title{Find the ancestral sequence at the MRCA of the tips from this host, or, if a dual infection was previously identified, of the MRCA of the tips making up each infection event}
\usage{
reconstruct.host.ancestral.sequences(phyloscanner.tree, host,
  individual.duals = F, verbose = F)
}
\arguments{
\item{phyloscanner.tree}{A list of class \code{phyloscanner.tree} (usually an item in a list of class \code{phyloscanner.trees}). This must have an \code{ancestral.alignment} element (see \emph{reconstruct.ancestral.sequences})}

\item{host}{The host ID}

\item{individual.duals}{Whether to output multiple sequences for \code{host} based on the results of a previous dual infection analysis}

\item{verbose}{Verbose output}
}
\description{
Find the ancestral sequence at the MRCA of the tips from this host, or, if a dual infection was previously identified, of the MRCA of the tips making up each infection event
}
