% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utility_functions.R
\name{select.windows.by.read.and.tip.count}
\alias{select.windows.by.read.and.tip.count}
\title{Select for further analysis relationship classifications by read and tip counts}
\usage{
\method{select}{windows.by.read.and.tip.count}(ptrees, dwin, tip.regex, min.reads, min.tips, verbose = F)
}
\arguments{
\item{ptrees}{A list of class \code{phyloscanner.trees} produced by \code{phyloscanner.analyse.trees}.}

\item{dwin}{A data frame produced by \code{classify.pairwise.relationships}.}

\item{tip.regex}{The regular expression used to identify host IDs in tip names}

\item{min.reads}{The minimum number of reads from a host in a window needed in order for that window to count in determining relationships involving that patient}

\item{min.tips}{The minimum number of tips from a host in a window needed in order for that window to count in determining relationships involving that patient}

\item{verbose}{Verbose output}
}
\value{
A data frame with viral phylogenetic classifications of pairwise host relationships in each deep sequence phylogeny
}
\description{
Select for further analysis relationship classifications by read and tip counts
}
\examples{
\dontrun{
require(phyloscannerR)
#
#	continue Rakai example,
#	load phyloscanner output from 'phyloscanner.analyse.trees'
#	
file	<- system.file(file.path('extdata','ptyr192_phsc_analyse_trees_output.R'),package='phyloscannerR')
load(file)	#loads 'phsc', output from 'phyloscanner.analyse.trees'
#	use distance thresholds found in analysis of Rakai couples
close.threshold <- 0.025
distant.threshold <- 0.05	
#	use relationship types based on adjacency
#	this also considers linkage etc between individuals who have dual infections, recombinants etc 
#	..and thus may not have *all* their subgraphs adjacent to each other
relationship.types <- c('proximity.3.way',					
		'close.and.adjacent',					
		'close.and.adjacent.and.directed',					
		'close.and.adjacent.and.ancestry.cat')	
dwin <- classify.pairwise.relationships(phsc, allow.mt=TRUE, close.threshold=close.threshold, distant.threshold=distant.threshold,relationship.types=relationship.types, verbose=TRUE)
tip.regex <- "^(.*)_fq[0-9]+_read_([0-9]+)_count_([0-9]+)$"
min.reads <- 30
min.tips <- 1
dwin <- select.windows.by.read.and.tip.count(phsc, dwin, tip.regex, min.reads, min.tips)
#
# 	end of Rakai example
#
}	 	
}
\author{
Oliver Ratmann, Matthew Hall
}
