% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utility_functions.R
\name{simmap.to.phyloscanner}
\alias{simmap.to.phyloscanner}
\title{Cast SIMMAP tree to phyloscanner tree}
\usage{
simmap.to.phyloscanner(ph, delete.simmap.structures)
}
\arguments{
\item{ph}{A \code{SIMMAP} tree with elements \code{maps}, \code{mapped.edge}, \code{node.states}.}

\item{delete.simmap.structures}{Logical value. If true \code{SIMMAP} elements are removed from the tree.}
}
\value{
Same tree in \code{phyloscanner.tree} format, with attributes \code{SPLIT}, \code{INDIVIDUAL}, \code{BRANCH_COLOURS}, \code{SUBGRAPH_MRCA}.
}
\description{
Cast SIMMAP tree to phyloscanner tree
}
\examples{
#
# Example on data from Rakai Community Cohort Study
#
require(phyloscannerR)
ph.phyloscanner <- readRDS(file=system.file(file.path('extdata','ptyr292_1600_to_1849.rds'),package='phyloscannerR'))

#	cast from phyloscanner format to SIMMAP format
#	so that the SIMMAP tree still has the phyloscanner attributes
ph.simmap <- phyloscanner.to.simmap(ph.phyloscanner, delete.phyloscanner.structures=FALSE)
attr(ph, "SPLIT")
#	to delete the phyloscanner attributes, use
ph.simmap <- phyloscanner.to.simmap(ph.phyloscanner, delete.phyloscanner.structures=TRUE)
#
#	cast from SIMMAP format to phyloscanner format
#	similarly, use delete.simmap.structures to keep/delete SIMMAP elements from the tree in phyloscanner format
ph.phyloscanner2 <- simmap.to.phyloscanner(ph.simmap, delete.simmap.structures=TRUE)
#
#	test that ph.phyloscanner2 is the same as ph.phyloscanner:
stopifnot(all( ph.phyloscanner$tip.label == ph.phyloscanner2$tip.label ))
stopifnot(all( ph.phyloscanner$edge == ph.phyloscanner2$edge ))
stopifnot(all( ph.phyloscanner$edge.length == ph.phyloscanner2$edge.length ))
stopifnot(all( ph.phyloscanner$Nnode == ph.phyloscanner2$Nnode ))
tmp <- as.character(attr(ph.phyloscanner,'SPLIT'))
tmp2 <- as.character(attr(ph.phyloscanner2,'SPLIT'))
tmp[is.na(tmp)] <- 'Unknown'
tmp2[is.na(tmp2)] <- 'Unknown'
stopifnot( all(tmp==tmp2) )
tmp <- as.character(attr(ph.phyloscanner,'INDIVIDUAL'))
tmp2 <- as.character(attr(ph.phyloscanner2,'INDIVIDUAL'))
tmp[is.na(tmp)] <- 'Unknown'
tmp2[is.na(tmp2)] <- 'Unknown'
stopifnot( all(tmp==tmp2) )
tmp <- as.character(attr(ph.phyloscanner,'BRANCH_COLOURS'))
tmp2 <- as.character(attr(ph.phyloscanner2,'BRANCH_COLOURS'))
tmp[is.na(tmp)] <- 'Unknown'
tmp2[is.na(tmp2)] <- 'Unknown'
stopifnot( all(tmp==tmp2) )
tmp <- attr(ph.phyloscanner,'SUBGRAPH_MRCA')
tmp2 <- attr(ph.phyloscanner2,'SUBGRAPH_MRCA')
stopifnot( all(tmp==tmp2) )
	
}
\seealso{
\code{\link{phyloscanner.to.simmap}}, \code{\link{extract.subgraph}}
}
\author{
Oliver Ratmann
}
