% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_analysis.R
\name{transmission.summary}
\alias{transmission.summary}
\title{Summarise the pairwise host relationships across all trees}
\usage{
transmission.summary(phyloscanner.trees, win.threshold = 0,
  dist.threshold = Inf, allow.mt = T, verbose = F)
}
\arguments{
\item{phyloscanner.trees}{A list of class \code{phyloscanner.trees}}

\item{win.threshold}{The proportion of windows that a pair of hosts need to be related (adjacent and within \code{dist.threshold} of each other) in order for them to appear in the summary.}

\item{dist.threshold}{The patristic distance within which the subgraphs from two hosts need to be in order for them to be declared related (default is infinity, so adjacent hosts are always related).}

\item{allow.mt}{If FALSE, directionality is only inferred between pairs of hosts where a single clade from one host is nested in one from the other; this is more conservative.}

\item{verbose}{Give verbose output}
}
\value{
A \code{data.table}, every line of which counts the number of pairwise relationships of a particular type between a pair of hosts
}
\description{
Summarise the pairwise host relationships across all trees
}
