% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utility_functions.R
\name{treeio.to.phyloscanner}
\alias{treeio.to.phyloscanner}
\title{Cast treeio tree to phyloscanner tree}
\usage{
treeio.to.phyloscanner(ph)
}
\arguments{
\item{ph}{A \code{treeio} tree with elements \code{phylo}, \code{data}. Element \code{data} is expected to have columns \code{node} and \code{location}.}
}
\value{
Same tree in \code{phyloscanner.tree} format, with attributes \code{SPLIT}, \code{INDIVIDUAL}, \code{BRANCH_COLOURS}, \code{SUBGRAPH_MRCA}.
}
\description{
Cast treeio tree to phyloscanner tree
}
\seealso{
\code{\link{phyloscanner.to.simmap}}, \code{\link{extract.subgraph}}
}
\author{
Oliver Ratmann
}
