% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_analysis.R
\name{write.annotated.tree}
\alias{write.annotated.tree}
\title{Write the phylogeny with reconstructed host annotations to file}
\usage{
write.annotated.tree(ptree, file.name, format = c("pdf", "nex"),
  pdf.scale.bar.width = 0.01, pdf.w = 50, pdf.hm = 0.15,
  verbose = F)
}
\arguments{
\item{file.name}{The name of the output file}

\item{format}{The format - PDF or NEXUS - in which to write the output.}

\item{pdf.scale.bar.width}{The width, in substitutions per site, of the scale bar in PDF output}

\item{pdf.w}{The width of the output PDF file, in inches}

\item{pdf.hm}{The height, in inches per tip, of the output PDF file}

\item{verbose}{Verbose output}

\item{phyloscanner.tree}{A list of class \code{phyloscanner.tree} (usually an item in a list of class \code{phyloscanner.trees})}
}
\description{
Write the phylogeny with reconstructed host annotations to file
}
