% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_summary_methods.R
\name{get.pairwise.relationships.any.ancestry.cat}
\alias{get.pairwise.relationships.any.ancestry.cat}
\title{Group to determine 2x2 table of distance and topology}
\usage{
get.pairwise.relationships.any.ancestry.cat(all.classifications)
}
\description{
Group to determine 2x2 table of distance and topology
}
\author{
Oliver Ratmann, Matthew Hall
}
\keyword{internal}
