% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_summary_methods.R
\name{get.pairwise.relationships.basic.by.ancestry}
\alias{get.pairwise.relationships.basic.by.ancestry}
\title{Define basic topological relationships between two individuals by ancestry}
\usage{
get.pairwise.relationships.basic.by.ancestry(all.classifications)
}
\description{
Define basic topological relationships between two individuals by ancestry
}
\author{
Oliver Ratmann
}
\keyword{internal}
