% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_summary_methods.R
\name{get.pairwise.relationships.close.and.contiguous.and.directed.cat}
\alias{get.pairwise.relationships.close.and.contiguous.and.directed.cat}
\title{Group to determine direction of transmission in likely pairs based on contiguous subgraphs}
\usage{
get.pairwise.relationships.close.and.contiguous.and.directed.cat(
  all.classifications
)
}
\description{
Group to determine direction of transmission in likely pairs based on contiguous subgraphs
}
\author{
Oliver Ratmann, Matthew Hall
}
\keyword{internal}
