% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.network}
\alias{plot.network}
\title{Plot transmission network}
\usage{
plot.network(df, di, control)
}
\arguments{
\item{df}{data.table with the following columns 'IDCLU','ID1', 'ID2', 'TYPE','KEFF','LKL_MAX','POSTERIOR_SCORE'}

\item{di}{data.table with meta-data to customize the plot with columns ID, node.shape, node.label, node.fill}

\item{control}{list of plotting control variables; see examples.}
}
\value{
ggplot object
}
\description{
This function plots a phylogenetic transmission network, showing three types of edges: 
two directed edges respectively in the 1->2 and 2->1 direction, and an undirected edge that represents phylogenetic support of close and adjacent individuals without evidence into the direction transmission.
}
\seealso{
\link{\code{find.networks}}
}
