% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{produce.pairwise.graphs}
\alias{produce.pairwise.graphs}
\title{Draw bar graphs of pairwise topological/distance relationships to file}
\usage{
produce.pairwise.graphs(ptrees, hosts = all.hosts.from.trees(ptrees),
  dist.thresh, out.file.name, width = 30, height.each = 5)
}
\arguments{
\item{ptrees}{A list of class \code{phyloscanner.trees}}

\item{hosts}{A list of hosts (as a vector) to obtain graphs for. By default, all pairs of hosts detected in \code{ptrees}.}

\item{width}{The width of the page of the output file in inches}

\item{file.name}{Output file name (expected to be a PDF)}

\item{height}{The height of each comparison in of the output file in inches}

\item{verbose}{Verbose output}
}
\description{
Draw bar graphs of pairwise topological/distance relationships to file
}
