% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{produce.pairwise.graphs}
\alias{produce.pairwise.graphs}
\title{Draw bar graphs of pairwise topological/distance relationships}
\usage{
produce.pairwise.graphs(ptrees, dist.thresh = 0.025,
  hosts = all.hosts.from.trees(ptrees), contiguous.pairs = F,
  inclusion = c("both", "either"))
}
\arguments{
\item{ptrees}{A list of class \code{phyloscanner.trees}}

\item{dist.thresh}{The distance threshold used to select likely transmission pairs}

\item{hosts}{A list of hosts (as a vector) to obtain graphs for. By default, all pairs of hosts detected in \code{ptrees}.}

\item{contiguous.pairs}{If TRUE pairs require contiguous (rather than ajacent) subgraphs to be identified as likely transmissions}

\item{inclusion}{If "both", then only pairs in which both individuals are members of \code{hosts} are included. If "either" then pairs only need have one member from \code{hosts}}
}
\value{
A list whose elements are \code{data}, the underlying data frame for the graph, and \code{graph}, the graph itself.
}
\description{
Draw bar graphs of pairwise topological/distance relationships
}
\examples{
#
# Example on data from Rakai Community Cohort Study
#
\dontrun{	
file	<- system.file(file.path('extdata','ptyr192_phsc_analyse_trees_output.RData'),package='phyloscannerR')
load(file)	#loads 'phsc', output from 'phyloscanner.analyse.trees'	
hosts	<- c('RkA05868F','RkA05968M','RkA00369F','RkA01344M')
inclusion <- "both" 
tmp	<- produce.pairwise.graphs(phsc, hosts=hosts, inclusion = "both")
tmp$graph
}
}
